Don't override global moretrees
This mod is currently overriding the global `moretrees` supplied by the respective mod.
This global is a necessary interface for other mods using moretrees. For example, installing hardtrees together with moretrees and plantlife will cause a crash because plantlife cannot access that interface anymore.
By simply making this variable local, this compatability issue is fixed.